requirements.txt
pyTelegramBotAPI==4.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
schedule==1.2.1
cryptography
psycopg2-binary>=2.9